Added: User Management, Roles & Permissions

+++ Modules/Product/Http/Controllers/ProductController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductDataTable;
 use Illuminate\Contracts\Support\Renderable;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Modules\Product\Entities\Product;
 use Modules\Product\Http\Requests\ProductCreateRequest;
@@ -15,27 +16,20 @@ use Modules\Upload\Entities\Upload;
 class ProductController extends Controller
 {
 
-    /**
-     * Display a listing of the resource.
-     * @return Renderable
-     */
     public function index(ProductDataTable $dataTable) {
+        abort_if(Gate::denies('access_products'), 403);
+
         return $dataTable->render('product::products.index');
     }
 
-    /**
-     * Show the form for creating a new resource.
-     * @return Renderable
-     */
+
     public function create() {
+        abort_if(Gate::denies('create_products'), 403);
+
         return view('product::products.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     * @param Request $request
-     * @return Renderable
-     */
+
     public function store(ProductCreateRequest $request) {
         $product = Product::create($request->except('image'));
 
@@ -55,38 +49,30 @@ class ProductController extends Controller
         return redirect()->route('products.index');
     }
 
-    /**
-     * Show the details for the specified resource.
-     * @param int $id
-     * @return Renderable
-     */
+
     public function show(Product $product) {
+        abort_if(Gate::denies('show_products'), 403);
+
         return view('product::products.show', compact('product'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     * @param int $id
-     * @return Renderable
-     */
+
     public function edit(Product $product) {
+        abort_if(Gate::denies('edit_products'), 403);
+
         return view('product::products.edit', compact('product'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     * @param Request $request
-     * @param int $id
-     * @return Renderable
-     */
+
     public function update(ProductUpdateRequest $request, Product $product) {
         $product->update($request->except('image'));
 
         if ($request->has('image')) {
             $tempFile = Upload::where('folder', $request->image)->first();
 
-            $media = $product->getMedia();
-            $media[0]->delete();
+            if ($product->getFirstMedia()) {
+                $product->getFirstMedia()->delete();
+            }
 
             if ($tempFile) {
                 $product->addMedia(Storage::path('public/temp/' . $request->image . '/' . $tempFile->filename))->toMediaCollection();
@@ -101,12 +87,10 @@ class ProductController extends Controller
         return redirect()->route('products.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     * @param int $id
-     * @return Renderable
-     */
+
     public function destroy(Product $product) {
+        abort_if(Gate::denies('delete_products'), 403);
+
         $product->delete();
 
         toast('Product Deleted!', 'warning');
